refactor(operator_brief): split seam persistence helpers
